requirements.txt
--extra-index-url https://mirrors.scut-smil.cn/nvidia/compute/redist
aiohttp
nvidia-dali-cuda110
pillow
